feat: move modules around by refactoring
