--- tests/data/context.yaml
@@ -1,14 +1,0 @@
-context:
-  name: "foo"
-  version: "1.2.3"
-  name_version: ${{ name }}-${{ version }}
-  cuda_version: ${{ hash }}_cuda_${{ version | version_to_buildstring }}
-  split_version: ${{ version | split('.') | join(';') }}
-
-package:
-  name: ${{ name }}
-  version: ${{ version }}
-  func_result: ${{ compiler('c') }}
-
-build:
-  string: ${{ blas_variant }}${{ hash }}_${{ name_version }}
